+++ app/retweet_graphs_v2/k_days/reporter.py
@@ -4,7 +4,7 @@ from pandas import DataFrame
 
 from app import DATA_DIR
 from app.retweet_graphs_v2.graph_storage import GraphStorage
-from app.retweet_graphs_v2.date_range_generator import DateRangeGenerator
+from app.retweet_graphs_v2.k_days.generator import DateRangeGenerator import DateRangeGenerator
 
 
 if __name__ == "__main__":
